Fix cert update command dropping to shell
The cert update command was dropping users to a shell instead of running
the installer. This was due to incorrect arguments passed to syscall.Exec.

The issue: syscall.Exec requires argv[0] to be the program name itself.
We were passing just the script path without 'bash' as the first argument.

Fix: Properly construct the arguments array with 'bash' as the first element
followed by the installer script path and any additional arguments.